chore: move rules file
